requirements.txt
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
idna==2.6
prettytable==0.7.2
protobuf==3.4.0
PyHive==0.5.1
pymongo==3.6.0
pymssql==2.1.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-toolbelt==0.8.0
six==1.10.0
socketIO-client-2==0.7.5
SQLAlchemy==1.2.3
urllib3==1.22
websocket-client==0.44.0
python-dateutil==2.6.1
PyMySQL==0.8.0
grpcio==1.10.0
grpcio-tools==1.10.0
python-consul==0.7.2
retrying==1.3.3
pyDes==2.0.1
